+++ src/pages/Editor/Description/DetailInformation/PlaceAndTimeEditor/index.ts
@@ -1,0 +1,1 @@
+export { PlaceAndTimeEditor } from "./PlaceAndTimeEditor";
